+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file. See [standa
 
 ## [Unreleased]
 
+### Fixed
+- Visual bug Integrations/ Select chatbot **[CHAT-8]**
+
 ### 0.0.16 (2025-06-26)
 
 ### Added