+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. See [standa
 
 ## [Unreleased]
 
+### Fixed
+
+- Documentation tab broken links
+
 ### 0.0.16 (2025-06-26)
 
 ### Added